+++ examples/platforms/dart/test/Properties_test.dart
@@ -21,6 +21,7 @@
 import "dart:typed_data";
 import "package:test/test.dart";
 import "package:hello/test.dart";
+import "package:hello/src/test/internal_class_with_static_property.dart";
 import "../test_suite.dart";
 
 final _testSuite = TestSuite("Properties");
@@ -89,5 +90,10 @@ void main() {
 
     expect(CachedProperties.staticCallCount, 1);
     expect(result1, Uint8List.fromList([0, 1, 2]));
+  });
+  _testSuite.test("Static internal property", () {
+    final result = InternalClassWithStaticProperty.internalfooBar;
+
+    expect(result, "foo");
   });
 }